requirements.txt
aiogram>=3.13.0
python-dotenv>=1.0.0
pytest>=8.0.0
httpx>=0.27.0
# PostgreSQL (Supabase и др.) — если задан DATABASE_URL
psycopg2-binary>=2.9.0
# Веб-интерфейс и авторизация
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
aiofiles>=24.1.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.12
apscheduler>=3.10.0
